+++ src/main/java/ee/carlrobert/llm/client/openai/completion/request/FileContext.java
@@ -1,0 +1,20 @@
+package ee.carlrobert.llm.client.openai.completion.request;
+
+public class FileContext {
+
+  private final String name;
+  private final String content;
+
+  public FileContext(String name, String content) {
+    this.name = name;
+    this.content = content;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public String getContent() {
+    return content;
+  }
+}